plugin duplicate files fix

+++ wp-content/plugins/seo-by-rank-math-pro/includes/admin/class-setup-wizard.php
@@ -116,8 +116,12 @@ class Setup_Wizard {
 
 		// Parse Options.
 		$wp_filesystem = WordPress::get_filesystem();
-		$settings      = $wp_filesystem->get_contents( $file['file'] );
-		$settings      = json_decode( $settings, true );
+		if ( is_null( $wp_filesystem ) ) {
+			return false;
+		}
+
+		$settings = $wp_filesystem->get_contents( $file['file'] );
+		$settings = json_decode( $settings, true );
 
 		\unlink( $file['file'] );
